\name{Arguments<-}
\alias{Arguments<-}
\title{Set the flydoc arguments documentation}
\usage{
  Arguments(func, value) <- value
}
\arguments{
  \item{func}{A function to add flydoc argument
  documentation to.}

  \item{value}{A named character vector containgin the
  argument documentation to add to the function's
  documentation}
}
\description{
  Set the documentation for the arguments for a given
  function
}

