\name{Description<-}
\alias{Description<-}
\title{Set a description}
\usage{
  Description(func, value) <- value
}
\arguments{
  \item{func}{A function to add a flydoc description to.}

  \item{value}{A character string containing the
  description to add to the function's documentation}
}
\description{
  Set a description for your flydoc documentation for a
  given function
}

