\name{Details<-}
\alias{Details<-}
\title{Set a detail section}
\usage{
  Details(func, value) <- value
}
\arguments{
  \item{func}{A function to add a flydoc details section
  to.}

  \item{value}{A character string containing the details
  section to add to the function's documentation}
}
\description{
  Set a title for your flydoc documentation for a given
  function
}

