\name{Examples<-}
\alias{Examples<-}
\title{Set some examples}
\usage{
  Examples(func, value) <- value
}
\arguments{
  \item{func}{A function to add examples to.}

  \item{value}{A character string containing the examples
  to add to the function's documentation}
}
\description{
  Set some examples for your flydoc documentation for a
  given function
}

