\docType{package}
\name{flydoc}
\alias{flydoc}
\alias{flydoc-package}
\alias{package-flydoc}
\title{Provide documentation to your functions on the fly}
\description{
  This package is a designed to provide an easy way to
  document your functions on the fly.  If you've defined a
  function interactively this package will still allow you
  to provide a help file.  If you have a function stored in
  a script but don't want to or have the time to create a
  package for it - this will allow you to still provide
  some help documentation for those functions.
}

