\name{Return<-}
\alias{Return<-}
\title{Set documentation for return}
\usage{
  Return(func, value) <- value
}
\arguments{
  \item{func}{A function to add a flydoc return to.}

  \item{value}{A character string containing the return to
  add to the function's documentation}
}
\description{
  Set a return for your flydoc documentation for a given
  function. This is what drive the section labeled "Value"
  in the help files.
}

