\name{Return}
\alias{Return}
\alias{Value}
\title{Look at flydoc return}
\usage{
  Return(fun)
}
\arguments{
  \item{fun}{A function}
}
\description{
  Look at the the return for your flydoc documnetation for
  a given function.  This is what drives the section
  labeled "Value" in the help files.
}

