\name{Title<-}
\alias{Title<-}
\title{Set a title}
\usage{
  Title(func, value) <- value
}
\arguments{
  \item{func}{A function to add a flydoc title to.}

  \item{value}{A character string containing the title to
  add to the function's documentation}
}
\description{
  Set a title for your flydoc documentation for a given
  function
}

